services/whisper/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
av==14.4.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.6.0
exceptiongroup==1.3.0
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.4
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnxruntime==1.22.0
packaging==25.0
pillow==11.0.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyreadline3==3.5.4
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
setuptools==80.9.0
sniffio==1.3.1
soundfile==0.13.1
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
websockets==15.0.1
